Move backorder update code to re-usable class
